Se desacopla las clases de Album sobre la vista general
