+++ flaskr/vistas/VistaAlbum.py
@@ -1,0 +1,67 @@
+import json
+from flaskr.modelos.modelos import AlbumSchema, Album
+from flaskr.modelos.modelos import Genero
+from flask import request
+from ..modelos import db, Cancion, Usuario
+from flask_restful import Resource
+from sqlalchemy.exc import IntegrityError
+from flask_jwt_extended import jwt_required
+
+album_schema = AlbumSchema()
+
+class VistaAlbum(Resource):
+    
+    def get(self, id_album):
+        return album_schema.dump(Album.query.get_or_404(id_album))
+
+    def put(self, id_album):
+        album = Album.query.get_or_404(id_album)
+        album.titulo = request.json.get("titulo",album.titulo)
+        album.anio = request.json.get("anio", album.anio)
+        album.descripcion = request.json.get("descripcion", album.descripcion)
+        album.medio = request.json.get("medio", album.medio)
+        db.session.commit()
+        return album_schema.dump(album)
+
+    def delete(self, id_album):
+        album = Album.query.get_or_404(id_album)
+        db.session.delete(album)
+        db.session.commit()
+        return '',204
+
+"""Se listan los generos @William Sanchez"""
+class VistaGenero(Resource):
+    def get(self): 
+        enum_list = list(map(lambda g: g.name, Genero))
+        return json.dumps(enum_list)
+
+
+class VistaAlbumesCanciones(Resource):
+    def get(self, id_cancion):
+        cancion = Cancion.query.get_or_404(id_cancion)
+        return [album_schema.dump(al) for al in cancion.albumes]
+
+"""Se agregan los generos @William Sanchez"""
+class VistaAlbumsUsuario(Resource):
+
+    @jwt_required()
+    def post(self, id_usuario):
+        nuevo_album = Album(titulo=request.json["titulo"], anio=request.json["anio"], descripcion=request.json["descripcion"], medio=request.json["medio"], genero=request.json["genero"])
+        usuario = Usuario.query.get_or_404(id_usuario)
+        usuario.albumes.append(nuevo_album)
+
+        try:
+            db.session.commit()
+        except IntegrityError:
+            db.session.rollback()
+            return 'El usuario ya tiene un album con dicho nombre',409
+
+        return album_schema.dump(nuevo_album)
+
+    @jwt_required()
+    def get(self, id_usuario):
+        usuario = Usuario.query.get_or_404(id_usuario)
+        return [album_schema.dump(al) for al in usuario.albumes]
+
+
+
