+++ flaskr/vistas/__init__.py
@@ -1,1 +1,2 @@
 from .vistas import *
+from .VistaAlbum import *
