+++ flaskr/vistas/vistas.py
@@ -1,5 +1,3 @@
-import json
-from flaskr.modelos.modelos import Genero
 from flask import request
 from ..modelos import db, Cancion, CancionSchema, Usuario, UsuarioSchema, Album, AlbumSchema
 from flask_restful import Resource
@@ -43,11 +41,6 @@ class VistaCancion(Resource):
         db.session.commit()
         return '',204
 
-class VistaAlbumesCanciones(Resource):
-    def get(self, id_cancion):
-        cancion = Cancion.query.get_or_404(id_cancion)
-        return [album_schema.dump(al) for al in cancion.albumes]
-
 class VistaSignIn(Resource):
     
     def post(self):
@@ -81,28 +74,6 @@ class VistaLogIn(Resource):
             token_de_acceso = create_access_token(identity = usuario.id)
             return {"mensaje":"Inicio de sesión exitoso", "token": token_de_acceso}
 
-"""Se agregan los generos @William Sanchez"""
-class VistaAlbumsUsuario(Resource):
-
-    @jwt_required()
-    def post(self, id_usuario):
-        nuevo_album = Album(titulo=request.json["titulo"], anio=request.json["anio"], descripcion=request.json["descripcion"], medio=request.json["medio"], genero=request.json["genero"])
-        usuario = Usuario.query.get_or_404(id_usuario)
-        usuario.albumes.append(nuevo_album)
-
-        try:
-            db.session.commit()
-        except IntegrityError:
-            db.session.rollback()
-            return 'El usuario ya tiene un album con dicho nombre',409
-
-        return album_schema.dump(nuevo_album)
-
-    @jwt_required()
-    def get(self, id_usuario):
-        usuario = Usuario.query.get_or_404(id_usuario)
-        return [album_schema.dump(al) for al in usuario.albumes]
-
 class VistaCancionesAlbum(Resource):
 
     def post(self, id_album):
@@ -126,28 +97,3 @@ class VistaCancionesAlbum(Resource):
         album = Album.query.get_or_404(id_album)
         return [cancion_schema.dump(ca) for ca in album.canciones]
 
-class VistaAlbum(Resource):
-
-    def get(self, id_album):
-        return album_schema.dump(Album.query.get_or_404(id_album))
-
-    def put(self, id_album):
-        album = Album.query.get_or_404(id_album)
-        album.titulo = request.json.get("titulo",album.titulo)
-        album.anio = request.json.get("anio", album.anio)
-        album.descripcion = request.json.get("descripcion", album.descripcion)
-        album.medio = request.json.get("medio", album.medio)
-        db.session.commit()
-        return album_schema.dump(album)
-
-    def delete(self, id_album):
-        album = Album.query.get_or_404(id_album)
-        db.session.delete(album)
-        db.session.commit()
-        return '',204
-
-"""Se listan los generos @William Sanchez"""
-class VistaGenero(Resource):
-    def get(self): 
-        enum_list = list(map(lambda g: g.name, Genero))
-        return json.dumps(enum_list)
